Refactor activity buckets logic into dedicated module
